fix: 修改获取用户特定小说阅读记录接口

+++ business/src/main/java/com/abdecd/novelbackend/business/pojo/dto/reader/DeleteReaderHistoryDTO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class DeleteReaderHistoryDTO {
     @NotNull
     @Schema(description = "历史记录ids")
-    Integer[] ids;
+    Long[] ids;
 }
